smartproctor/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic[email-validator]==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
aiofiles==24.1.0
email-validator==2.1.1